Created locationData.js file in seeds and copy/pasted over data from locations.js.  Commented out locations.  Minor bug fixes
