+++ seeds/userData.json
@@ -13,7 +13,7 @@
       "name": "James",
       "email": "James@aol.com",
       "password": "password12345"
-    }
+    },
     {
         "name": "Beau",
         "email": "Beau@aol.com",
